Move registering Omegaconf resolvers to `utils.config` module
